Rename project to not include api
