+++ src/ImmersiveReaderDemo/ImmersiveReaderDemo.csproj
