+++ src/ImmersiveReaderDemo/appsettings.Development.json
